registrySACHER_backend/requirements.txt
Flask==0.12.2
psycopg2==2.7.4
toml==0.9.4
python-cinderclient==3.4.0
python-glanceclient==2.11.1
python-keystoneclient==3.17.0
python-magic==0.4.15
python-novaclient==9.1.1
python-openstackclient==3.13.0
python-swiftclient==3.0.0
paramiko